Add some different logics in logic class
